Go: selection sort
