+++ Sortings/SelectionSort/Python/Analysis/best_case.py
