+++ Sortings/SelectionSort/Python/Analysis/worst_case.py
